Refactor: Move data layer classes to core.data package.
This commit relocates classes related to the data layer, including entities, DAOs, models, repositories, and enums, from the `com.currency.rateman.data` package to `com.currency.rateman.core.data`.

This change aims to improve the project structure by centralizing core data components. Corresponding import statements in other files have been updated to reflect this move.
