+++ app/src/main/java/com/currency/rateman/api/kurzyCz/ProviderConverter.kt
@@ -1,12 +1,12 @@
 package com.currency.rateman.api.kurzyCz
 
 import android.util.Log
-import com.currency.rateman.data.db.entity.CurrencyRateEntity
+import com.currency.rateman.core.data.db.entity.CurrencyRateEntity
 import com.currency.rateman.provider.data.entity.ProviderEntity
-import com.currency.rateman.data.model.CurrencyRate
+import com.currency.rateman.core.data.model.CurrencyRate
 import com.currency.rateman.provider.data.model.Provider
-import com.currency.rateman.data.model.enums.CurrencyCode
-import com.currency.rateman.data.model.enums.ProviderType
+import com.currency.rateman.core.data.model.enums.CurrencyCode
+import com.currency.rateman.core.data.model.enums.ProviderType
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
