+++ app/src/main/java/com/currency/rateman/api/scraper/fetchAuraAktivRates.kt
@@ -1,6 +1,6 @@
 package com.currency.rateman.api.scraper
 
-import com.currency.rateman.data.model.enums.CurrencyCode
+import com.currency.rateman.core.data.model.enums.CurrencyCode
 import com.currency.rateman.provider.utils.ProviderConstants
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
