+++ app/src/main/java/com/currency/rateman/core/data/model/CurrencyRate.kt
@@ -1,6 +1,6 @@
-package com.currency.rateman.data.model
+package com.currency.rateman.core.data.model
 
-import com.currency.rateman.data.model.enums.CurrencyCode
+import com.currency.rateman.core.data.model.enums.CurrencyCode
 import java.time.LocalDate
 
 data class CurrencyRate(
