+++ app/src/main/java/com/currency/rateman/core/data/model/Filter.kt
@@ -1,9 +1,9 @@
-package com.currency.rateman.data.model
+package com.currency.rateman.core.data.model
 
 import android.os.Parcelable
-import com.currency.rateman.data.model.enums.CurrencyCode
-import com.currency.rateman.data.model.enums.ProviderType
-import com.currency.rateman.data.model.enums.RateSortType
+import com.currency.rateman.core.data.model.enums.CurrencyCode
+import com.currency.rateman.core.data.model.enums.ProviderType
+import com.currency.rateman.core.data.model.enums.RateSortType
 import kotlinx.parcelize.Parcelize
 import kotlinx.serialization.Serializable
