+++ app/src/main/java/com/currency/rateman/core/data/model/Settings.kt
@@ -1,9 +1,9 @@
-package com.currency.rateman.data.model
+package com.currency.rateman.core.data.model
 
 import android.os.Parcelable
-import com.currency.rateman.data.model.enums.CurrencyCode
-import com.currency.rateman.data.model.enums.LanguageCode
-import com.currency.rateman.data.model.enums.ThemeMode
+import com.currency.rateman.core.data.model.enums.CurrencyCode
+import com.currency.rateman.core.data.model.enums.LanguageCode
+import com.currency.rateman.core.data.model.enums.ThemeMode
 import kotlinx.parcelize.Parcelize
 import kotlinx.serialization.Serializable
