+++ app/src/main/java/com/currency/rateman/core/data/model/enums/LanguageCode.kt
@@ -1,4 +1,4 @@
-package com.currency.rateman.data.model.enums
+package com.currency.rateman.core.data.model.enums
 
 enum class LanguageCode {
     EN, CS, RU, ES, UK, KY, TR, SK
