+++ app/src/main/java/com/currency/rateman/core/data/model/enums/ProviderType.kt
@@ -1,4 +1,4 @@
-package com.currency.rateman.data.model.enums
+package com.currency.rateman.core.data.model.enums
 
 enum class ProviderType {
     BANK, EXCHANGE, CRYPTO_EXCHANGE, ALL
