+++ app/src/main/java/com/currency/rateman/core/data/model/enums/RateSortType.kt
@@ -1,4 +1,4 @@
-package com.currency.rateman.data.model.enums
+package com.currency.rateman.core.data.model.enums
 
 enum class RateSortType {
     BEST_RATE, BEST_BUY, BEST_SELL
