+++ app/src/main/java/com/currency/rateman/core/data/model/enums/ThemeMode.kt
@@ -1,0 +1,5 @@
+package com.currency.rateman.core.data.model.enums
+
+enum class ThemeMode {
+    DARK, LIGHT
+}
