+++ app/src/main/java/com/currency/rateman/core/data/repository/SettingsRepository.kt
@@ -1,9 +1,9 @@
-package com.currency.rateman.data.repository
+package com.currency.rateman.core.data.repository
 
-import com.currency.rateman.data.model.enums.CurrencyCode
-import com.currency.rateman.data.model.enums.LanguageCode
-import com.currency.rateman.data.model.Settings
-import com.currency.rateman.data.model.enums.ThemeMode
+import com.currency.rateman.core.data.model.Settings
+import com.currency.rateman.core.data.model.enums.CurrencyCode
+import com.currency.rateman.core.data.model.enums.LanguageCode
+import com.currency.rateman.core.data.model.enums.ThemeMode
 import kotlinx.coroutines.flow.Flow
 
 interface SettingsRepository {
