--- app/src/main/java/com/currency/rateman/data/model/enums/ThemeMode.kt
@@ -1,5 +1,0 @@
-package com.currency.rateman.data.model.enums
-
-enum class ThemeMode {
-    DARK, LIGHT
-}
