+++ app/src/main/java/com/currency/rateman/provider/data/model/Provider.kt
@@ -1,8 +1,8 @@
 package com.currency.rateman.provider.data.model
 
-import com.currency.rateman.data.model.CurrencyRate
-import com.currency.rateman.data.model.enums.CurrencyCode
-import com.currency.rateman.data.model.enums.ProviderType
+import com.currency.rateman.core.data.model.CurrencyRate
+import com.currency.rateman.core.data.model.enums.CurrencyCode
+import com.currency.rateman.core.data.model.enums.ProviderType
 
 data class Provider(
     val id: Long,
