+++ app/src/main/java/com/currency/rateman/provider/ui/components/ProviderItem.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
-import com.currency.rateman.data.model.CurrencyRate
+import com.currency.rateman.core.data.model.CurrencyRate
 import com.currency.rateman.provider.data.model.Provider
 import com.currency.rateman.utils.formatRate
