+++ app/src/main/java/com/currency/rateman/provider/ui/viewmodel/ProviderDetailViewModel.kt
@@ -2,7 +2,7 @@ package com.currency.rateman.provider.ui.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.currency.rateman.data.model.enums.CurrencyCode
+import com.currency.rateman.core.data.model.enums.CurrencyCode
 import com.currency.rateman.provider.data.model.Provider
 import com.currency.rateman.provider.domain.repository.ProviderRepository
 import kotlinx.coroutines.flow.MutableStateFlow
