+++ app/src/main/java/com/currency/rateman/ui/theme/Theme.kt
@@ -8,7 +8,7 @@ import androidx.compose.runtime.collectAsState
 import com.currency.rateman.ui.viewmodels.SettingsViewModel
 import org.koin.androidx.compose.koinViewModel
 import androidx.compose.runtime.getValue
-import com.currency.rateman.data.model.enums.ThemeMode
+import com.currency.rateman.core.data.model.enums.ThemeMode
 
 private val DarkColorScheme = darkColorScheme(
     primary = DarkPrimary,
