+++ app/src/main/java/com/currency/rateman/ui/viewmodels/SettingsViewModel.kt
@@ -3,11 +3,11 @@ package com.currency.rateman.ui.viewmodels
 import android.content.Context
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.currency.rateman.data.model.Settings
-import com.currency.rateman.data.model.enums.CurrencyCode
-import com.currency.rateman.data.model.enums.LanguageCode
-import com.currency.rateman.data.model.enums.ThemeMode
-import com.currency.rateman.data.repository.SettingsRepository
+import com.currency.rateman.core.data.model.Settings
+import com.currency.rateman.core.data.model.enums.CurrencyCode
+import com.currency.rateman.core.data.model.enums.LanguageCode
+import com.currency.rateman.core.data.model.enums.ThemeMode
+import com.currency.rateman.core.data.repository.SettingsRepository
 import com.currency.rateman.utils.LanguageManager
 import com.currency.rateman.utils.ThemeManager
 import kotlinx.coroutines.flow.MutableStateFlow
