+++ app/src/main/java/com/currency/rateman/utils/LanguageManager.kt
@@ -4,10 +4,10 @@ import android.content.Context
 import android.content.res.Configuration
 import androidx.appcompat.app.AppCompatDelegate
 import androidx.core.os.LocaleListCompat
-import com.currency.rateman.data.db.RateManDatabase
-import com.currency.rateman.data.model.enums.LanguageCode
-import com.currency.rateman.data.repository.SettingsRepository
-import com.currency.rateman.data.repository.SettingsRepositoryImpl
+import com.currency.rateman.core.data.db.RateManDatabase
+import com.currency.rateman.core.data.model.enums.LanguageCode
+import com.currency.rateman.core.data.repository.SettingsRepository
+import com.currency.rateman.core.data.repository.SettingsRepositoryImpl
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.runBlocking
 import java.util.*
